+++ src/main/kotlin/vn/loi/learning/domain/study/session/model/PracticeLoopPolicy.kt
@@ -2,5 +2,7 @@ package vn.loi.learning.domain.study.session.model
 
 enum class PracticeLoopPolicy {
     NONE,
-    LOOP_FIXED_MEMBERSHIP_SHUFFLED
+    LOOP_FIXED_MEMBERSHIP_SHUFFLED,
+    LOOP_ADAPTIVE_FEEDBACK_SHUFFLED,
+    LOOP_DYNAMIC_DIFFICULT_MEMBERSHIP
 }